chore: move `compose()` code
